+++ src/components/Header/Header.css
@@ -6,10 +6,11 @@
 header {
   z-index: 3;
   position: fixed;
+  top: 0;
   display: flex;
   background: rgb(183, 153, 215);
   width: 100vw;
-  height: 5vh;
+  height: 10vh;
   margin-top: -10px;
   margin-left: -10px;
   box-shadow: rgba(0, 0, 0, 0.16) 0px 3px 6px, rgba(0, 0, 0, 0.23) 0px 3px 6px;
